* master:
  ja.lproj  --> Japanese.lproj etc. within .xcodeproj
  cleaning up files
  minor fixes
  the simplified Chinese localization should now work.
  base localization. somehow zh-Hans doesn't get compiled to binary. Hmm.
  moved to base localization. fingers crossed
  minor adjustments
  Added support for:
  Update README.md
  Fix over-release
  Remove memory leaks

# Conflicts:
#	PrefPane/Base.lproj/MenuMetersPref.xib
#	PrefPane/English.lproj/MenuMetersPref.nib/keyedobjects.nib

+++ MenuExtras/MenuMeterDisk/MenuMeterDiskSpace.m
@@ -198,7 +198,7 @@ static NSComparisonResult SortDiskEntryByDeviceString(NSDictionary *a, NSDiction
 						  forKey:@"free"];
 			[diskStats setObject:[NSString stringWithFormat:[localizedStrings objectForKey:kUsedSpaceFormat],
 									[self spaceString:(((float)mountInfo[i].f_blocks -
-														(float)mountInfo[i].f_bavail) * mountInfo[i].f_bsize)]]
+														(float)mountInfo[i].f_bavail) * (float)mountInfo[i].f_bsize)]]
 						  forKey:@"used"];
 			// Store the data into the array
 			[diskSpaceDetails addObject:diskStats];
